BeforeGraf

+++ Assets/Scripts/ScoreCounter.cs
@@ -7,13 +7,13 @@ public class ScoreCounter : MonoBehaviour
 {
     [SerializeField] private Text scoreText;
 
-    public float score = 0;
+    public int score = 0;
     void Start()
     {
         scoreText = GameObject.Find("Score").GetComponent<Text>();
     }
 
-    public void AddPoint(float inc)
+    public void AddPoint(int inc)
     {
         score += inc;
         scoreText.text = score.ToString();
